+++ qcs-api-client-grpc/src/tonic/refresh.rs
@@ -4,6 +4,7 @@ use std::{
     task::{Context, Poll},
 };
 
+use http::StatusCode;
 use tonic::{
     body::BoxBody,
     client::GrpcService,
@@ -12,12 +13,6 @@ use tonic::{
 use tower::Layer;
 
 use qcs_api_client_common::configuration::{ClientConfiguration, TokenError, TokenRefresher};
-
-#[cfg(feature = "tracing-opentelemetry")]
-use super::traced_service_call;
-
-#[cfg(not(feature = "tracing-opentelemetry"))]
-use super::service_call;
 
 use super::error::Error;
 
@@ -114,17 +109,77 @@ where
         // https://github.com/tower-rs/tower/issues/547
         let service = std::mem::replace(&mut self.service, service);
         let token_refresher = self.token_refresher.clone();
-        Box::pin(async move {
-            #[cfg(feature = "tracing-opentelemetry")]
-            use opentelemetry_api::trace::FutureExt;
-
-            #[cfg(feature = "tracing-opentelemetry")]
-            return traced_service_call(req, token_refresher, service)
-                .with_current_context()
-                .await;
-
-            #[cfg(not(feature = "tracing-opentelemetry"))]
-            return service_call(req, token_refresher, service).await;
-        })
+        super::common::pin_future_with_otel_context_if_available(service_call(
+            req,
+            token_refresher,
+            service,
+        ))
     }
+}
+
+async fn service_call<C, T>(
+    req: Request<BoxBody>,
+    token_refresher: T,
+    mut channel: C,
+) -> Result<Response<<C as GrpcService<BoxBody>>::ResponseBody>, Error<T::Error>>
+where
+    C: GrpcService<BoxBody> + Send,
+    <C as GrpcService<BoxBody>>::ResponseBody: Send,
+    <C as GrpcService<BoxBody>>::Future: Send,
+    T: TokenRefresher + Send,
+    T::Error: std::error::Error,
+    Error<T::Error>: From<C::Error>,
+{
+    let token = token_refresher
+        .validated_access_token()
+        .await
+        .map_err(Error::Refresh)?;
+    let (req, retry_req) = super::build_duplicate_request(req).await?;
+    let resp = make_request(&mut channel, req, token).await?;
+
+    let grpc_authnz_failure = matches!(
+        super::common::get_status_code_from_headers(resp.headers()).ok(),
+        Some(tonic::Code::Unauthenticated) | Some(tonic::Code::PermissionDenied)
+    );
+    let http_authnz_failure =
+        resp.status() == StatusCode::UNAUTHORIZED || resp.status() == StatusCode::FORBIDDEN;
+
+    if grpc_authnz_failure || http_authnz_failure {
+        #[cfg(feature = "tracing")]
+        {
+            tracing::info!("refreshing token after receiving unauthorized or forbidden status",);
+        }
+
+        // Refresh token and try again
+        let token = token_refresher
+            .validated_access_token()
+            .await
+            .map_err(Error::Refresh)?;
+
+        #[cfg(feature = "tracing")]
+        {
+            tracing::info!("token refreshed");
+        }
+        make_request(&mut channel, retry_req, token).await
+    } else {
+        Ok(resp)
+    }
+}
+
+async fn make_request<C, E: std::error::Error>(
+    service: &mut C,
+    mut request: Request<BoxBody>,
+    token: String,
+) -> Result<Response<<C as GrpcService<BoxBody>>::ResponseBody>, Error<E>>
+where
+    C: GrpcService<BoxBody> + Send,
+    <C as GrpcService<BoxBody>>::ResponseBody: Send,
+    <C as GrpcService<BoxBody>>::Future: Send,
+    Error<E>: From<C::Error>,
+{
+    let header_val = format!("Bearer {token}")
+        .try_into()
+        .map_err(Error::InvalidAccessToken)?;
+    request.headers_mut().insert("authorization", header_val);
+    service.call(request).await.map_err(Error::from)
 }